requirements.txt
python-dotenv==1.0.1
aiohttp==3.10.11
